+++ tests/test_triggers.py
@@ -9,7 +9,8 @@ from rakaia.callbacks.triggers import (
     no_channel_for_view,
     empty_slider_values,
     use_channel_autofill,
-    layout_has_modified_shape)
+    layout_has_modified_shape,
+    stitch_cache_delete)
 import numpy as np
 
 def test_triggers():
@@ -52,6 +53,10 @@ def test_triggers():
     assert not use_channel_autofill(rgb_layers, "roi_1", "channel_1", True, True)
     assert not use_channel_autofill(rgb_layers, "roi_1", "channel_2", False, True)
     assert not use_channel_autofill(rgb_layers, "roi_1", "channel_2", True, False)
+
+    assert not stitch_cache_delete("wrong_id", "stitch_1")
+    assert not stitch_cache_delete("stitch-image-delete", None)
+    assert stitch_cache_delete("stitch-image-delete", "to_delete")
 
 
 def test_annotation_index_triggers():